Create Feedback.java
I tried using JFrame at first, but this seems to be a much simpler solution.

When the route is finished, we can call the function getFeedback() to display a pop that shows a message with yes/no/cancel response button. It will return an integer (0 if they hit yes, 1 if they hit no, and 3 if they hit cancel). Can that be easily transferred to your database?

I'll see if I can find the API to check when the route is finished (since we don't want to get the feedback until they've arrived at their destination).